Added Functionality for Changing Labels
A common use case is when the user accidentally selects the wrong label.
In prior versions, the only way to change a label was to go into the
database and update labels manually. In this new version, you can update
labels using the `taskStats-detail` page. We now have a dropdown box for
all the labels for the current task, and you can change the label and
press `Update` to save.